requirements.txt
openai>=1.0.0
httpx>=0.23.0
pytesseract
pillow
mss
keyboard
pyyaml
googletrans==4.0.0-rc1